refactor(DTO): Restructure SiteDto/SiteVerboseDto
