+++ bpdm-bridge-dummy/src/main/kotlin/com/catenax/bpdm/bridge/dummy/service/PoolUpdateService.kt
@@ -23,9 +23,9 @@ import com.catenax.bpdm.bridge.dummy.dto.*
 import mu.KotlinLogging
 import org.eclipse.tractusx.bpdm.common.dto.BusinessPartnerType
 import org.eclipse.tractusx.bpdm.common.dto.LogisticAddressDto
-import org.eclipse.tractusx.bpdm.common.dto.SiteDto
 import org.eclipse.tractusx.bpdm.gate.api.client.GateClient
 import org.eclipse.tractusx.bpdm.pool.api.client.PoolApiClient
+import org.eclipse.tractusx.bpdm.pool.api.model.SiteDto
 import org.eclipse.tractusx.bpdm.pool.api.model.request.*
 import org.eclipse.tractusx.bpdm.pool.api.model.response.*
 import org.springframework.stereotype.Service
@@ -75,7 +75,7 @@ class PoolUpdateService(
                     SitePartnerCreateRequest(
                         site = SiteDto(
                             name = entry.site.nameParts.firstOrNull() ?: "",
-                            states = entry.site.states,
+                            states = entry.site.states.map(::gateToPoolSiteState),
                             mainAddress = gateToPoolLogisticAddress(entry.mainAddress),
                         ),
                         index = entry.externalId,
@@ -99,7 +99,7 @@ class PoolUpdateService(
             SitePartnerUpdateRequest(
                 site = SiteDto(
                     name = it.site.nameParts.firstOrNull() ?: "",
-                    states = it.site.states,
+                    states = it.site.states.map(::gateToPoolSiteState),
                     mainAddress = gateToPoolLogisticAddress(it.mainAddress),
                 ),
                 bpns = it.bpn!!
