+++ bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/SiteGateDto.kt
@@ -22,9 +22,10 @@ package org.eclipse.tractusx.bpdm.gate.api.model
 import io.swagger.v3.oas.annotations.media.ArraySchema
 import io.swagger.v3.oas.annotations.media.Schema
 import org.eclipse.tractusx.bpdm.common.dto.BusinessPartnerRole
-import org.eclipse.tractusx.bpdm.common.dto.SiteStateDto
 import org.eclipse.tractusx.bpdm.common.dto.openapidescription.SiteDescription
 
+// TODO: This DTO does not implement IBaseSiteDto because it is missing a [name] property.
+//  Removing this from IBaseSiteDto has lots of implications. And this DTO is only used in a deprecated API anyways!
 @Schema(description = SiteDescription.header)
 data class SiteGateDto(
