+++ bpdm-pool-api/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/api/model/SiteStateDto.kt
@@ -17,27 +17,20 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.eclipse.tractusx.bpdm.common.dto.response
+package org.eclipse.tractusx.bpdm.pool.api.model
 
 import io.swagger.v3.oas.annotations.media.Schema
+import org.eclipse.tractusx.bpdm.common.dto.ISiteStateDto
 import org.eclipse.tractusx.bpdm.common.dto.openapidescription.SiteStateDescription
-import org.eclipse.tractusx.bpdm.common.dto.response.type.TypeKeyNameVerboseDto
 import org.eclipse.tractusx.bpdm.common.model.BusinessStateType
 import java.time.LocalDateTime
 
 @Schema(description = SiteStateDescription.header)
-data class SiteStateVerboseDto(
+data class SiteStateDto(
 
-    @get:Schema(description = SiteStateDescription.description)
-    val description: String?,
+    override val description: String?,
+    override val validFrom: LocalDateTime?,
+    override val validTo: LocalDateTime?,
+    override val type: BusinessStateType
 
-    @get:Schema(description = SiteStateDescription.validFrom)
-    val validFrom: LocalDateTime?,
-
-    @get:Schema(description = SiteStateDescription.validTo)
-    val validTo: LocalDateTime?,
-
-    // TODO OpenAPI description for complex field does not work!!
-    @get:Schema(description = SiteStateDescription.type)
-    val type: TypeKeyNameVerboseDto<BusinessStateType>
-)
+) : ISiteStateDto
