+++ bpdm-pool-api/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/api/model/SiteStateVerboseDto.kt
@@ -17,32 +17,27 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.eclipse.tractusx.bpdm.common.dto.response
+package org.eclipse.tractusx.bpdm.pool.api.model
 
-import io.swagger.v3.oas.annotations.media.ArraySchema
 import io.swagger.v3.oas.annotations.media.Schema
-import org.eclipse.tractusx.bpdm.common.dto.openapidescription.CommonDescription
-import org.eclipse.tractusx.bpdm.common.dto.openapidescription.SiteDescription
-import java.time.Instant
+import org.eclipse.tractusx.bpdm.common.dto.openapidescription.SiteStateDescription
+import org.eclipse.tractusx.bpdm.common.dto.response.type.TypeKeyNameVerboseDto
+import org.eclipse.tractusx.bpdm.common.model.BusinessStateType
+import java.time.LocalDateTime
 
-@Schema(description = SiteDescription.header)
-data class SiteVerboseDto(
+@Schema(description = SiteStateDescription.header)
+data class SiteStateVerboseDto(
 
-    @get:Schema(description = SiteDescription.bpns)
-    val bpns: String,
+    @get:Schema(description = SiteStateDescription.description)
+    val description: String?,
 
-    @get:Schema(description = SiteDescription.name)
-    val name: String,
+    @get:Schema(description = SiteStateDescription.validFrom)
+    val validFrom: LocalDateTime?,
 
-    @ArraySchema(arraySchema = Schema(description = SiteDescription.states))
-    val states: Collection<SiteStateVerboseDto> = emptyList(),
+    @get:Schema(description = SiteStateDescription.validTo)
+    val validTo: LocalDateTime?,
 
-    @get:Schema(description = SiteDescription.bpnLegalEntity)
-    val bpnLegalEntity: String,
-
-    @get:Schema(description = CommonDescription.createdAt)
-    val createdAt: Instant,
-
-    @get:Schema(description = CommonDescription.updatedAt)
-    val updatedAt: Instant
+    // TODO OpenAPI description for complex field does not work!!
+    @get:Schema(description = SiteStateDescription.type)
+    val type: TypeKeyNameVerboseDto<BusinessStateType>
 )
