+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/ResponseMappings.kt
@@ -223,8 +223,8 @@ fun Site.toDto(): SiteVerboseDto {
     )
 }
 
-fun Site.toPoolDto(): SitePoolVerboseDto {
-    return SitePoolVerboseDto(
+fun Site.toPoolDto(): SiteWithMainAddressVerboseDto {
+    return SiteWithMainAddressVerboseDto(
 
         site = SiteVerboseDto(
             bpn,
